% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesixtygiving-package.R
\docType{package}
\name{threesixtygiving}
\alias{threesixtygiving-package}
\title{threesixtygiving package}
\description{
Charity funding data for the UK. See the
[360 Giving website](https://www.threesixtygiving.org) for more information
on the source of this data.
}
