% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-grants.R
\name{tsg_all_grants}
\alias{tsg_all_grants}
\title{All grants}
\usage{
tsg_all_grants(verbose = TRUE, timeout = 30, retries = 3)
}
\arguments{
\item{verbose}{If `TRUE`, prints console messages.}

\item{timeout}{The maximum request time, in seconds. If data is not returned
in this time, a value of `NA` is returned for that dataset.}

\item{retries}{The number of retries to make if a request is not successful.
Defaults to 3.}
}
\value{
A list of tibbles with grant data.
}
\description{
Returns a list of data frames with details of all grants from funders
returned by [tsg_available()].
}
\details{
Due to the structure of the 360 Giving data, the package will make
multiple attempts to request data. This can cause issues with servers
automatically blocking requests, you can use [tsg_missing()] to identify
missing sets of grant data.
}
\examples{
\donttest{
all_grants <- tsg_all_grants()
}

}
