% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-data.R
\name{tsg_missing}
\alias{tsg_missing}
\title{Identify missing data sets}
\usage{
tsg_missing(x)
}
\arguments{
\item{x}{A list of tibble with grant data returned by \code{\link[=tsg_all_grants]{tsg_all_grants()}}.}
}
\value{
A tibble with details on funders missing data from a list of tibbles
passed to the function.
}
\description{
Returns a tibble with details on any available data missing from a list
of grants returned by \code{\link[=tsg_all_grants]{tsg_all_grants()}}.
}
\examples{
\dontrun{
all_grants <- tsg_all_grants()

missing_grants <- tsg_missing(all_grants)
}

}
