% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{tsg_process_data}
\alias{tsg_process_data}
\title{Process to tibble}
\usage{
tsg_process_data(x, verbose = TRUE)
}
\arguments{
\item{x}{A list of tibble with grant data returned by [tsg_all_grants()].}

\item{verbose}{If `TRUE`, prints console messages.}
}
\value{
A tibble with all variables from all grants
}
\description{
Given a list returned by [tsg_all_grants()] or [tsg_search_grants()], creates
a tibble with all available variables. All variables are converted to
character class. This tibble contains roughly 200 columns if all available
grants are used (as of October 2019), due to differences in labelling and
structuring grant data.
}
\examples{
\donttest{
grants <- tsg_all_grants()

df <- tsg_process_data(grants)
}
}
\seealso{
tsg_core_data
}
