% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-data.R
\name{tsg_specific_data}
\alias{tsg_specific_data}
\title{Retrieve specific datasets}
\usage{
tsg_specific_data(search, search_in = NULL, verbose = TRUE,
  ignore_case = TRUE, perl = FALSE, fixed = FALSE)
}
\arguments{
\item{search}{The string(s) to search for. By default allows POSIX 1003.2
regular expressions. Use `perl = TRUE` for perl-style regex.
Accepts single strings or a character vector.}

\item{search_in}{The name of the column to search in. Accepts single strings
or a character vector of column names. If `NULL`, searches all columns.}

\item{verbose}{If `TRUE`, prints console messages.}

\item{ignore_case}{If `TRUE` ignores case.}

\item{perl}{If `TRUE`, uses perl-style regex.}

\item{fixed}{If `TRUE`, searches will be matched as-is.}
}
\value{
A single tibble (if only one grant maker matches the queries) or a
list of tibbles (if the query matches multiple datasets).
}
\description{
Returns a list of tibble with details of all grants from funders
where funder data matches one or more search strings. If only one dataset
matches queries, returns a tibble of that dataset.
}
\examples{
\donttest{

}
}
