% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-df.R
\name{tsg_specific_df}
\alias{tsg_specific_df}
\title{Get specific datasets}
\usage{
tsg_specific_df(x, verbose = TRUE, timeout = 30, retries = 3)
}
\arguments{
\item{x}{All or a subset of a tibble returned by
[tsg_available()] or [tsg_missing()]}

\item{verbose}{If `TRUE`, prints console messages.}

\item{timeout}{The maximum request time, in seconds. If data is not returned
in this time, a value of `NA` is returned for that dataset.}

\item{retries}{The number of retries to make if a request is not successful.
Defaults to 3.}
}
\value{
A list of tibbles with grant data.
}
\description{
Retrieve data contained with all or specific rows of a tibble
returned by [tsg_available()] or [tsg_missing()].
}
\examples{
\donttest{
all_grants <- tsg_all_grants()

missing_grants <- tsg_missing(all_grants)

more_grants <- tsg_specific_df(missing_grants)
}

}
\seealso{
tsg_specific_data_search

tsg_available

tsg_missing
}
