% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threesixtygiving-package.R
\docType{package}
\name{threesixtygiving}
\alias{threesixtygiving-package}
\title{threesixtygiving package}
\description{
360Giving is a data standard for publishing information about
charitable grant giving in the UK. \code{threesixtygiving} provides functions to
retrieve and process charitable grant funding data from 360Giving.
}
\details{
See the \href{https://www.threesixtygiving.org/}{360 Giving website}
for more information on the source of this data, and the
\href{https://standard.threesixtygiving.org}{360Giving Standard} page for
details of the data standard this package relies on.
}
