% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-grants.R
\name{tsg_all_grants}
\alias{tsg_all_grants}
\title{All grants}
\usage{
tsg_all_grants(core_data = FALSE, verbose = TRUE)
}
\arguments{
\item{core_data}{If `TRUE`}

\item{verbose}{If `TRUE`, prints console messages.}
}
\value{
A list of data frames.
}
\description{
Returns a list of data frames with details of all grants from funders
returned by [tsg_available()].
}
\examples{
\donttest{
all_grants <- tsg_all_grants()
}
}
