% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-data.R
\name{tsg_available}
\alias{tsg_available}
\title{Available datasets}
\usage{
tsg_available()
}
\value{
A tibble with details on all available datasets.
}
\description{
Information on the available datasets
}
\examples{
\dontrun{
available <- tsg_available()
}

}
