% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{tsg_process_data}
\alias{tsg_process_data}
\title{Convert data to tibble}
\usage{
tsg_process_data(x, min_coverage = 0, verbose = TRUE)
}
\arguments{
\item{x}{A list of tibble with grant data returned by [tsg_all_grants()].}

\item{min_coverage}{A number from 0 to 1. If >0, only returns variables
with a value other than `NA` for at least that proportion of rows. Defaults
to 0 and returns all columns}

\item{verbose}{If `TRUE`, prints console messages.}
}
\value{
A tibble with all variables from all grants.
}
\description{
Given a list returned by [tsg_all_grants()] or [tsg_search_grants()],
creates a tibble with all available variables. This tibble contains roughly
200 columns if all
available grants are used (as of October 2019), due to differences in
how grant data is labelled and structured.
}
\examples{
\dontrun{
grants <- tsg_all_grants()

df <- tsg_process_data(grants)
}

}
\seealso{
tsg_core_data
}
